+++ src/Entity/npc.h
@@ -76,9 +76,9 @@ public:
 
 public:
     /* constructor */
-    Npc(int32_t aUID, const char* aName,
+    Npc(uint32_t aUID, const char* aName,
         uint8_t aType, int16_t aLook,
-        int16_t aMapId, uint16_t aPosX, uint16_t aPosY,
+        uint32_t aMapId, uint16_t aPosX, uint16_t aPosY,
         uint8_t aBase, uint8_t aSort);
 
     /* destructor */
